# docs/openapi.yaml
components:
   schemas:
...
      Category:
         properties:
            description:
               anyOf:
               -  type: string
               -  type: 'null'
               title: Description
            id:
               title: Id
               type: integer
            is_active:
               default: true
               title: Is Active
               type: boolean
            name:
               title: Name
               type: string
         required:
         - id
         - name
         title: Category
         type: object
      CategoryCreateRequest:
         properties:
            description:
               anyOf:
               -  type: string
               -  type: 'null'
               title: Description
            is_active:
               default: true
               title: Is Active
               type: boolean
            name:
               title: Name
               type: string
         required:
         - name
         title: CategoryCreateRequest
         type: object
      CategoryUpdateRequest:
         properties:
            description:
               anyOf:
               -  type: string
               -  type: 'null'
               title: Description
            is_active:
               anyOf:
               -  type: boolean
               -  type: 'null'
               title: Is Active
            name:
               anyOf:
               -  type: string
               -  type: 'null'
               title: Name
         title: CategoryUpdateRequest
         type: object
      HTTPValidationError:
         properties:
            detail:
               items:
                  $ref: '#/components/schemas/ValidationError'
               title: Detail
               type: array
         title: HTTPValidationError
         type: object
...
      ValidationError:
         properties:
            loc:
               items:
                  anyOf:
                  -  type: string
                  -  type: integer
               title: Location
               type: array
            msg:
               title: Message
               type: string
            type:
               title: Error Type
               type: string
         required:
         - loc
         - msg
         - type
         title: ValidationError
         type: object
info:
   description: "\u0420\u0435\u0444\u0430\u043A\u0442\u043E\u0440\u0438\u0440\u043E\
      \u0432\u0430\u043D\u043D\u044B\u0439 API \u0434\u043B\u044F \u0441\u0438\u0441\
      \u0442\u0435\u043C\u044B \u0443\u043F\u0440\u0430\u0432\u043B\u0435\u043D\u0438\
      \u044F \u043A\u043E\u043B\u043B\u0430\u0436\u0430\u043C\u0438"
   title: CollagrCms API
   version: 0.1.0
openapi: 3.1.0
paths:
   /categories/:
      get:
         description: "\u041F\u043E\u043B\u0443\u0447\u0430\u0435\u0442 \u0432\u0441\
            \u0435 \u043A\u0430\u0442\u0435\u0433\u043E\u0440\u0438\u0438 (\u0432\u043A\
            \u043B\u044E\u0447\u0430\u044F \u043D\u0435\u0430\u043A\u0442\u0438\u0432\
            \u043D\u044B\u0435)\n\nReturns:\n    \u0421\u043F\u0438\u0441\u043E\u043A\
            \ \u0432\u0441\u0435\u0445 \u043A\u0430\u0442\u0435\u0433\u043E\u0440\u0438\
            \u0439"
         operationId: get_categories_categories__get
         responses:
            '200':
               content:
                  application/json:
                     schema:
                        items:
                           $ref: '#/components/schemas/Category'
                        title: Response Get Categories Categories  Get
                        type: array
               description: Successful Response
         summary: Get Categories
         tags:
         - categories
      post:
         description: "\u0421\u043E\u0437\u0434\u0430\u0435\u0442 \u043D\u043E\u0432\
            \u0443\u044E \u043A\u0430\u0442\u0435\u0433\u043E\u0440\u0438\u044E\n\n\
            Args:\n    request: \u0414\u0430\u043D\u043D\u044B\u0435 \u0434\u043B\u044F\
            \ \u0441\u043E\u0437\u0434\u0430\u043D\u0438\u044F \u043A\u0430\u0442\u0435\
            \u0433\u043E\u0440\u0438\u0438\n\nReturns:\n    \u0421\u043E\u0437\u0434\
            \u0430\u043D\u043D\u0430\u044F \u043A\u0430\u0442\u0435\u0433\u043E\u0440\
            \u0438\u044F"
         operationId: create_category_categories__post
         requestBody:
            content:
               application/json:
                  schema:
                     $ref: '#/components/schemas/CategoryCreateRequest'
            required: true
         responses:
            '201':
               content:
                  application/json:
                     schema:
                        $ref: '#/components/schemas/Category'
               description: Successful Response
            '422':
               content:
                  application/json:
                     schema:
                        $ref: '#/components/schemas/HTTPValidationError'
               description: Validation Error
         summary: Create Category
         tags:
         - categories
...
   /categories/{category_id}:
      delete:
         description: "\u0423\u0434\u0430\u043B\u044F\u0435\u0442 \u043A\u0430\u0442\
            \u0435\u0433\u043E\u0440\u0438\u044E (\u043C\u044F\u0433\u043A\u043E\u0435\
            \ \u0443\u0434\u0430\u043B\u0435\u043D\u0438\u0435 - \u0434\u0435\u0430\
            \u043A\u0442\u0438\u0432\u0438\u0440\u0443\u0435\u0442)\n\nArgs:\n   \
            \ category_id: ID \u043A\u0430\u0442\u0435\u0433\u043E\u0440\u0438\u0438\
            \n\nReturns:\n    \u0421\u043E\u043E\u0431\u0449\u0435\u043D\u0438\u0435\
            \ \u043E\u0431 \u0443\u0441\u043F\u0435\u0448\u043D\u043E\u043C \u0443\
            \u0434\u0430\u043B\u0435\u043D\u0438\u0438"
         operationId: delete_category_categories__category_id__delete
         parameters:
         -  in: path
            name: category_id
            required: true
            schema:
               title: Category Id
               type: integer
         responses:
            '200':
               content:
                  application/json:
                     schema: {}
               description: Successful Response
            '422':
               content:
                  application/json:
                     schema:
                        $ref: '#/components/schemas/HTTPValidationError'
               description: Validation Error
         summary: Delete Category
         tags:
         - categories
      get:
         description: "\u041F\u043E\u043B\u0443\u0447\u0430\u0435\u0442 \u043A\u0430\
            \u0442\u0435\u0433\u043E\u0440\u0438\u044E \u043F\u043E ID\n\nArgs:\n\
            \    category_id: ID \u043A\u0430\u0442\u0435\u0433\u043E\u0440\u0438\u0438\
            \n\nReturns:\n    \u041A\u0430\u0442\u0435\u0433\u043E\u0440\u0438\u044F"
         operationId: get_category_categories__category_id__get
         parameters:
         -  in: path
            name: category_id
            required: true
            schema:
               title: Category Id
               type: integer
         responses:
            '200':
               content:
                  application/json:
                     schema:
                        $ref: '#/components/schemas/Category'
               description: Successful Response
            '422':
               content:
                  application/json:
                     schema:
                        $ref: '#/components/schemas/HTTPValidationError'
               description: Validation Error
         summary: Get Category
         tags:
         - categories
      put:
         description: "\u041E\u0431\u043D\u043E\u0432\u043B\u044F\u0435\u0442 \u043A\
            \u0430\u0442\u0435\u0433\u043E\u0440\u0438\u044E\n\nArgs:\n    category_id:\
            \ ID \u043A\u0430\u0442\u0435\u0433\u043E\u0440\u0438\u0438\n    request:\
            \ \u0414\u0430\u043D\u043D\u044B\u0435 \u0434\u043B\u044F \u043E\u0431\
            \u043D\u043E\u0432\u043B\u0435\u043D\u0438\u044F \u043A\u0430\u0442\u0435\
            \u0433\u043E\u0440\u0438\u0438\n\nReturns:\n    \u041E\u0431\u043D\u043E\
            \u0432\u043B\u0435\u043D\u043D\u0430\u044F \u043A\u0430\u0442\u0435\u0433\
            \u043E\u0440\u0438\u044F"
         operationId: update_category_categories__category_id__put
         parameters:
         -  in: path
            name: category_id
            required: true
            schema:
               title: Category Id
               type: integer
         requestBody:
            content:
               application/json:
                  schema:
                     $ref: '#/components/schemas/CategoryUpdateRequest'
            required: true
         responses:
            '200':
               content:
                  application/json:
                     schema:
                        $ref: '#/components/schemas/Category'
               description: Successful Response
            '422':
               content:
                  application/json:
                     schema:
                        $ref: '#/components/schemas/HTTPValidationError'
               description: Validation Error
         summary: Update Category
         tags:
         - categories
   /categories/{category_id}/hard:
      delete:
         description: "\u041F\u043E\u043B\u043D\u043E\u0441\u0442\u044C\u044E \u0443\
            \u0434\u0430\u043B\u044F\u0435\u0442 \u043A\u0430\u0442\u0435\u0433\u043E\
            \u0440\u0438\u044E \u0438\u0437 \u0431\u0430\u0437\u044B \u0434\u0430\u043D\
            \u043D\u044B\u0445\n\nArgs:\n    category_id: ID \u043A\u0430\u0442\u0435\
            \u0433\u043E\u0440\u0438\u0438\n\nReturns:\n    \u0421\u043E\u043E\u0431\
            \u0449\u0435\u043D\u0438\u0435 \u043E\u0431 \u0443\u0441\u043F\u0435\u0448\
            \u043D\u043E\u043C \u0443\u0434\u0430\u043B\u0435\u043D\u0438\u0438"
         operationId: hard_delete_category_categories__category_id__hard_delete
         parameters:
         -  in: path
            name: category_id
            required: true
            schema:
               title: Category Id
               type: integer
         responses:
            '200':
               content:
                  application/json:
                     schema: {}
               description: Successful Response
            '422':
               content:
                  application/json:
                     schema:
                        $ref: '#/components/schemas/HTTPValidationError'
               description: Validation Error
         summary: Hard Delete Category
         tags:
         - categories
...